pyproject.toml
[project]
name = "multi-database-agent"
version = "1.0.0"
description = "Production-ready multi-database agent supporting MongoDB and Oracle via MCP servers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.14.1",
    "python-dotenv>=1.1.1",
    "cx-oracle>=8.3.0",
    "pymongo>=4.6.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "streamlit>=1.28.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.0",
    "oracledb>=1.4.0",
]
